+++ lib/src/commands/jayree/org/configure/country.d.ts
@@ -1,6 +1,8 @@
-import { flags, SfdxCommand } from '@salesforce/command';
+import { flags } from '@salesforce/command';
 import { AnyJson } from '@salesforce/ts-types';
-export default class UpdateCountry extends SfdxCommand {
+import { JayreeSfdxCommand } from '../../../../jayreeSfdxCommand';
+export default class UpdateCountry extends JayreeSfdxCommand {
+    static aliases: string[];
     static description: string;
     protected static flagsConfig: {
         silent: flags.Discriminated<flags.Boolean<boolean>>;