+++ lib/src/commands/jayree/source/retrieve/full.d.ts
@@ -3,6 +3,7 @@ import { AnyJson } from '@salesforce/ts-types';
 import { SourceRetrieveBase } from '../../../../sourceRetrieveBase';
 export default class RetrieveProfiles extends SourceRetrieveBase {
     static description: string;
+    static hidden: boolean;
     protected static flagsConfig: {
         keepcache: flags.Discriminated<flags.Boolean<boolean>>;
         metadata: flags.Discriminated<flags.Array<string>>;