+++ lib/src/jayreeSfdxCommand.d.ts
@@ -1,0 +1,4 @@
+import { SfdxCommand } from '@salesforce/command';
+export declare abstract class JayreeSfdxCommand extends SfdxCommand {
+    warnIfRunByAlias(commandClass: any): void;
+}