+++ src/commands/jayree/org/configure/country.ts
@@ -4,18 +4,21 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { core, flags, SfdxCommand } from '@salesforce/command';
+import { core, flags } from '@salesforce/command';
 import { AnyJson } from '@salesforce/ts-types';
 import chalk from 'chalk';
 import { cli } from 'cli-ux';
 import puppeteer from 'puppeteer';
 import { Tabletojson as tabletojson } from 'tabletojson';
-import * as config from '../../../../../config/countrystate.json';
+import config from '../../../../utils/config';
+import * as CSconfig from '../../../../../config/countrystate.json';
+import { JayreeSfdxCommand } from '../../../../jayreeSfdxCommand';
 
 core.Messages.importMessagesDirectory(__dirname);
 const messages = core.Messages.loadMessages('sfdx-jayree', 'createstatecountry');
 
-export default class UpdateCountry extends SfdxCommand {
+export default class UpdateCountry extends JayreeSfdxCommand {
+  public static aliases = ['jayree:automation:country:update'];
   public static description = messages.getMessage('commandCountryDescription');
 
   protected static flagsConfig = {
@@ -32,11 +35,10 @@ export default class UpdateCountry extends SfdxCommand {
   protected static requiresProject = false;
 
   public async run(): Promise<AnyJson> {
+    this.warnIfRunByAlias(UpdateCountry);
     let spinnermessage = '';
 
-    const browser = await puppeteer.launch({
-      headless: true,
-    });
+    const browser = await puppeteer.launch(config().puppeteer);
 
     const page = await browser.newPage();
 
@@ -128,7 +130,7 @@ export default class UpdateCountry extends SfdxCommand {
             waitUntil: 'networkidle0',
           }
         );
-        const setCountrySelector = config.setCountry;
+        const setCountrySelector = CSconfig.setCountry;
         await setHTMLInputElementValue(countryCode, setCountrySelector.editIntVal);
 
         await page.click(setCountrySelector.save.replace(/:/g, '\\:'));