+++ src/commands/jayree/source/retrieve/full.ts
@@ -22,6 +22,7 @@ const messages = core.Messages.loadMessages('sfdx-jayree', 'sourceretrievefull')
 
 export default class RetrieveProfiles extends SourceRetrieveBase {
   public static description = messages.getMessage('commandDescription');
+  public static hidden = true;
 
   /*   public static examples = [
     `$ sfdx jayree:flowtestcoverage
@@ -54,6 +55,7 @@ Coverage: 82%
   protected static requiresProject = true;
 
   public async run(): Promise<AnyJson> {
+    this.ux.warn('You are using a deprecated command. See the Readme file for more information.');
     await this.org.refreshAuth();
 
     const projectpath = this.project.getPath();
@@ -76,11 +78,10 @@ Coverage: 82%
       let packagexml = path.join(__dirname, '..', '..', '..', '..', '..', '..', 'manifest', 'package-profiles.xml');
 
       const pjson = await xml2js.parseStringPromise(fs.readFileSync(packagexml, 'utf8'));
-      pjson.Package.types[
-        pjson.Package.types.findIndex((x) => x.name.toString() === 'CustomObject')
-      ].members = pjson.Package.types[
-        pjson.Package.types.findIndex((x) => x.name.toString() === 'CustomObject')
-      ].members.concat(config(this.project.getPath()).ensureObjectPermissions);
+      pjson.Package.types[pjson.Package.types.findIndex((x) => x.name.toString() === 'CustomObject')].members =
+        pjson.Package.types[pjson.Package.types.findIndex((x) => x.name.toString() === 'CustomObject')].members.concat(
+          config(this.project.getPath()).ensureObjectPermissions
+        );
 
       packagexml = path.join(orgretrievepath, 'pinject.xml');
       await fs.writeFile(packagexml, builder.buildObject(pjson));