+++ test/commands/jayree/openorg.test.ts
@@ -18,7 +18,7 @@ describe('platform.win32', () => {
   });
 
   after(function () {
-    // restore original process.platfork
+    // restore original process.platform
     Object.defineProperty(process, 'platform', this.originalPlatform);
   });
   test
@@ -51,30 +51,6 @@ describe('platform.win32', () => {
     });
 });
 
-describe('platform.undefined', () => {
-  before(function () {
-    // save original process.platform
-    this.originalPlatform = Object.getOwnPropertyDescriptor(process, 'platform');
-
-    // redefine process.platform
-    Object.defineProperty(process, 'platform', {
-      value: 'undefined',
-    });
-  });
-
-  after(function () {
-    // restore original process.platfork
-    Object.defineProperty(process, 'platform', this.originalPlatform);
-  });
-  test
-    .withOrg({ username: 'test@org.com' }, true)
-    .stderr()
-    .command(['jayree:org:open', '--targetusername', 'test@org.com', '-r'])
-    .it('runs jayree:org:open --targetusername test@org.com -r', (ctx) => {
-      expect(ctx.stderr).to.contain('not supported');
-    });
-});
-
 describe('platform.linux', () => {
   before(function () {
     // save original process.platform
@@ -87,7 +63,7 @@ describe('platform.linux', () => {
   });
 
   after(function () {
-    // restore original process.platfork
+    // restore original process.platform
     Object.defineProperty(process, 'platform', this.originalPlatform);
   });
   test
@@ -132,7 +108,7 @@ describe('platform.darwin', () => {
   });
 
   after(function () {
-    // restore original process.platfork
+    // restore original process.platform
     Object.defineProperty(process, 'platform', this.originalPlatform);
   });
   test